Add mutex protection to Libft containers

+++ Template/queue.hpp
@@ -5,6 +5,8 @@
 #include "../CMA/CMA.hpp"
 #include "../Errno/errno.hpp"
 #include "../CPP_class/nullptr.hpp"
+#include "../Libft/libft.hpp"
+#include "../PThread/mutex.hpp"
 #include <cstddef>
 #include <utility>
 
@@ -18,10 +20,11 @@ class ft_queue
     	    QueueNode* next;
     	};
 
-    	QueueNode*  _front;
-    	QueueNode*  _rear;
-    	size_t      _size;
-    	mutable int _errorCode;
+        QueueNode*  _front;
+        QueueNode*  _rear;
+        size_t      _size;
+        mutable int _errorCode;
+        mutable pt_mutex _mutex;
 
     	void    setError(int error) const;
 
@@ -35,20 +38,20 @@ class ft_queue
     	ft_queue(ft_queue&& other) noexcept;
     	ft_queue& operator=(ft_queue&& other) noexcept;
 
-    	void enqueue(const ElementType& value);
-    	void enqueue(ElementType&& value);
-    	ElementType dequeue();
+        void enqueue(const ElementType& value);
+        void enqueue(ElementType&& value);
+        ElementType dequeue();
 
-    	ElementType& front();
-    	const ElementType& front() const;
+        ElementType& front();
+        const ElementType& front() const;
 
-    	size_t size() const;
-    	bool empty() const;
+        size_t size() const;
+        bool empty() const;
 
-    	int get_error() const;
-    	const char* get_error_str() const;
+        int get_error() const;
+        const char* get_error_str() const;
 
-    	void clear();
+        void clear();
 };
 
 template <typename ElementType>
@@ -61,7 +64,7 @@ ft_queue<ElementType>::ft_queue()
 template <typename ElementType>
 ft_queue<ElementType>::~ft_queue()
 {
-    clear();
+    this->clear();
     return ;
 }
 
@@ -81,15 +84,24 @@ ft_queue<ElementType>& ft_queue<ElementType>::operator=(ft_queue&& other) noexce
 {
     if (this != &other)
     {
-        clear();
-        _front = other._front;
-        _rear = other._rear;
-        _size = other._size;
-        _errorCode = other._errorCode;
+        if (this->_mutex.lock(THREAD_ID) != SUCCES)
+            return (*this);
+        if (other._mutex.lock(THREAD_ID) != SUCCES)
+        {
+            this->_mutex.unlock(THREAD_ID);
+            return (*this);
+        }
+        this->clear();
+        this->_front = other._front;
+        this->_rear = other._rear;
+        this->_size = other._size;
+        this->_errorCode = other._errorCode;
         other._front = ft_nullptr;
         other._rear = ft_nullptr;
         other._size = 0;
         other._errorCode = ER_SUCCESS;
+        other._mutex.unlock(THREAD_ID);
+        this->_mutex.unlock(THREAD_ID);
     }
     return (*this);
 }
@@ -97,7 +109,7 @@ ft_queue<ElementType>& ft_queue<ElementType>::operator=(ft_queue&& other) noexce
 template <typename ElementType>
 void ft_queue<ElementType>::setError(int error) const
 {
-    _errorCode = error;
+    this->_errorCode = error;
     ft_errno = error;
     return ;
 }
@@ -105,69 +117,90 @@ void ft_queue<ElementType>::setError(int error) const
 template <typename ElementType>
 void ft_queue<ElementType>::enqueue(const ElementType& value)
 {
+    if (this->_mutex.lock(THREAD_ID) != SUCCES)
+    {
+        this->setError(PT_ERR_MUTEX_OWNER);
+        return ;
+    }
     QueueNode* node = static_cast<QueueNode*>(cma_malloc(sizeof(QueueNode)));
     if (node == ft_nullptr)
     {
-        setError(QUEUE_ALLOC_FAIL);
+        this->setError(QUEUE_ALLOC_FAIL);
+        this->_mutex.unlock(THREAD_ID);
         return ;
     }
     construct_at(&node->data, value);
     node->next = ft_nullptr;
-    if (_rear == ft_nullptr)
+    if (this->_rear == ft_nullptr)
     {
-        _front = node;
-        _rear = node;
+        this->_front = node;
+        this->_rear = node;
     }
     else
     {
-        _rear->next = node;
-        _rear = node;
+        this->_rear->next = node;
+        this->_rear = node;
     }
-    ++_size;
+    ++this->_size;
+    this->_mutex.unlock(THREAD_ID);
     return ;
 }
 
 template <typename ElementType>
 void ft_queue<ElementType>::enqueue(ElementType&& value)
 {
+    if (this->_mutex.lock(THREAD_ID) != SUCCES)
+    {
+        this->setError(PT_ERR_MUTEX_OWNER);
+        return ;
+    }
     QueueNode* node = static_cast<QueueNode*>(cma_malloc(sizeof(QueueNode)));
     if (node == ft_nullptr)
     {
-        setError(QUEUE_ALLOC_FAIL);
+        this->setError(QUEUE_ALLOC_FAIL);
+        this->_mutex.unlock(THREAD_ID);
         return ;
     }
     construct_at(&node->data, std::move(value));
     node->next = ft_nullptr;
-    if (_rear == ft_nullptr)
+    if (this->_rear == ft_nullptr)
     {
-        _front = node;
-        _rear = node;
+        this->_front = node;
+        this->_rear = node;
     }
     else
     {
-        _rear->next = node;
-        _rear = node;
+        this->_rear->next = node;
+        this->_rear = node;
     }
-    ++_size;
+    ++this->_size;
+    this->_mutex.unlock(THREAD_ID);
     return ;
 }
 
 template <typename ElementType>
 ElementType ft_queue<ElementType>::dequeue()
 {
-    if (_front == ft_nullptr)
+    if (this->_mutex.lock(THREAD_ID) != SUCCES)
     {
-        setError(QUEUE_EMPTY);
+        this->setError(PT_ERR_MUTEX_OWNER);
         return (ElementType());
     }
-    QueueNode* node = _front;
-    _front = node->next;
-    if (_front == ft_nullptr)
-        _rear = ft_nullptr;
+    if (this->_front == ft_nullptr)
+    {
+        this->setError(QUEUE_EMPTY);
+        this->_mutex.unlock(THREAD_ID);
+        return (ElementType());
+    }
+    QueueNode* node = this->_front;
+    this->_front = node->next;
+    if (this->_front == ft_nullptr)
+        this->_rear = ft_nullptr;
     ElementType value = std::move(node->data);
     destroy_at(&node->data);
     cma_free(node);
-    --_size;
+    --this->_size;
+    this->_mutex.unlock(THREAD_ID);
     return (value);
 }
 
@@ -175,62 +208,97 @@ template <typename ElementType>
 ElementType& ft_queue<ElementType>::front()
 {
     static ElementType errorElement = ElementType();
-    if (_front == ft_nullptr)
+    if (this->_mutex.lock(THREAD_ID) != SUCCES)
     {
-        setError(QUEUE_EMPTY);
+        this->setError(PT_ERR_MUTEX_OWNER);
         return (errorElement);
     }
-    return (_front->data);
+    if (this->_front == ft_nullptr)
+    {
+        this->setError(QUEUE_EMPTY);
+        this->_mutex.unlock(THREAD_ID);
+        return (errorElement);
+    }
+    ElementType& ref = this->_front->data;
+    this->_mutex.unlock(THREAD_ID);
+    return (ref);
 }
 
 template <typename ElementType>
 const ElementType& ft_queue<ElementType>::front() const
 {
     static ElementType errorElement = ElementType();
-    if (_front == ft_nullptr)
+    if (this->_mutex.lock(THREAD_ID) != SUCCES)
     {
-        setError(QUEUE_EMPTY);
+        this->setError(PT_ERR_MUTEX_OWNER);
         return (errorElement);
     }
-    return (_front->data);
+    if (this->_front == ft_nullptr)
+    {
+        this->setError(QUEUE_EMPTY);
+        this->_mutex.unlock(THREAD_ID);
+        return (errorElement);
+    }
+    ElementType& ref = this->_front->data;
+    this->_mutex.unlock(THREAD_ID);
+    return (ref);
 }
 
 template <typename ElementType>
 size_t ft_queue<ElementType>::size() const
 {
-    return (_size);
+    if (this->_mutex.lock(THREAD_ID) != SUCCES)
+        return (0);
+    size_t s = this->_size;
+    this->_mutex.unlock(THREAD_ID);
+    return (s);
 }
 
 template <typename ElementType>
 bool ft_queue<ElementType>::empty() const
 {
-    return (_size == 0);
+    if (this->_mutex.lock(THREAD_ID) != SUCCES)
+        return (true);
+    bool res = (this->_size == 0);
+    this->_mutex.unlock(THREAD_ID);
+    return (res);
 }
 
 template <typename ElementType>
 int ft_queue<ElementType>::get_error() const
 {
-    return (_errorCode);
+    if (this->_mutex.lock(THREAD_ID) != SUCCES)
+        return (this->_errorCode);
+    int err = this->_errorCode;
+    this->_mutex.unlock(THREAD_ID);
+    return (err);
 }
 
 template <typename ElementType>
 const char* ft_queue<ElementType>::get_error_str() const
 {
-    return (ft_strerror(_errorCode));
+    if (this->_mutex.lock(THREAD_ID) != SUCCES)
+        return (ft_strerror(this->_errorCode));
+    int err = this->_errorCode;
+    this->_mutex.unlock(THREAD_ID);
+    return (ft_strerror(err));
 }
 
 template <typename ElementType>
 void ft_queue<ElementType>::clear()
 {
-    while (_front != ft_nullptr)
+    if (this->_mutex.lock(THREAD_ID) != SUCCES)
+        return ;
+    while (this->_front != ft_nullptr)
     {
-        QueueNode* node = _front;
-        _front = _front->next;
+        QueueNode* node = this->_front;
+        this->_front = this->_front->next;
         destroy_at(&node->data);
         cma_free(node);
     }
-    _rear = ft_nullptr;
-    _size = 0;
+    this->_rear = ft_nullptr;
+    this->_size = 0;
+    this->_mutex.unlock(THREAD_ID);
     return ;
 }
